+++ src/main.rs
@@ -12,4 +12,3 @@ mod tests;
 
 
 fn main() {}
-
